backend/requirements.txt
Django==3.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
Pillow==9.3.0
PyJWT==2.1.0
requests==2.26.0
django-colorfield==0.9.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.1
drf-extra-fields==3.7.0
flake8==5.0.4
